refactor(traces): fns/helpers.ts becomes three files named after what they do (LFE-14804)
  helpers.ts -> nestObservations.ts  (git mv, so the history follows the biggest
                                      export, plus its private getObservationLevels)
             -> heatMapTextColor.ts
             -> getSubtreeDurationOverflowMs.ts

One export per file, each named after it — rules 3 and 4 both. The private
helper stays in the file with the export that uses it, which the RFC allows and
which is why this is three files and not four.

rule 3: 13 -> 12, rule 4: 8 -> 7, traces total 134 -> 132.
